Rename actions, refactor code
