+++ Apps.Marketo/Models/Emails/Requests/ListEmailsRequest.cs
@@ -6,6 +6,7 @@ namespace Apps.Marketo.Models.Emails.Requests;
 
 public class ListEmailsRequest
 {
+    [DataSource(typeof(StatusDataHandler))]
     public string? Status { get; set; }
 
     [Display("Folder")]
@@ -16,7 +17,9 @@ public class ListEmailsRequest
     [DataSource(typeof(FolderTypeDataHandler))]
     public string? Type { get; set; }
 
-    [Display("Max return (maximum 200)")]
+    [Display("Earliest updated at")]
     public DateTime? EarliestUpdatedAt { get; set; }
+
+    [Display("Latest updated at")]
     public DateTime? LatestUpdatedAt { get; set; }
 }
